Move constant for buggy passes
