+++ ecmascript/minifier/src/compress/optimize/conditionals.rs
@@ -1,10 +1,10 @@
 use super::Optimizer;
 use crate::compress::optimize::bools::negate_cost;
 use crate::compress::optimize::Ctx;
-use crate::compress::optimize::DISABLE_BUGGY_PASSES;
 use crate::debug::dump;
 use crate::util::make_bool;
 use crate::util::SpanExt;
+use crate::DISABLE_BUGGY_PASSES;
 use std::mem::swap;
 use swc_common::EqIgnoreSpan;
 use swc_common::Spanned;
